Added Node Cluster
nodejs\Dockerfile installs nodejs and npm within the directory usr/src/app on port 3000 and runs app.js when an instance of a container is built.

nginx/nginx.conf - distributes traffic to 127.17.0.1:81-83 and listens on port 80

nginx\Dockerfile pulls nginx image then removes the default configuration file from the image and replaces it with the custom one in nginx\nginx.conf

docker-compose.yaml - uses version 3 and lists each container which uses the node.js dockerfile and listens in ports 81-83:3000 and builds dockerfile from nginx folder which exposes port 80:80 which distributes traffic to the conatiners. This doesn't build till all nodes are built.