Added Code For MongoDB using SweetShopDB For Now
nodejs\src\app.js - modified code to add a connection to Mongodb and creating a stock schema for the sweetshop database, with post and get methods testing and working in the rest client.

nodejs\Dockerfile - now installs express and mongoose and pm-2 globally using the pm2-runtime to run the app.js file

docker-compose.yaml - added a cluster of three mongo containers, on port 40001-40003:27017

scripts\mongo_setup - sleeps for ten seconds to allow cluster to build, then attaches to mongo1 and runs the configuration which specifies the three nodes, and mongo1 has a higher priority than the other nodes.