Change Hostnames and Create Node Message as JavaScript Object
Added Hostname for each app container so its easy to identify which one is which.

Added ability to get host name, created a list of nodes which are assigned a random id number.

Established a message which is sent by each node containing the hostname, nodeID and time.  The time is in seconds, and will be used to track how long it has been since the last message sent.