Subscriber to populate node list.
Subscriber code which receives a message from a node, and checks if it is currently in the nodeList. If it does exist then it just updates the node with the current value of 'seconds' to show the time of last message, if not it pushes the incoming node to the array.