Added Alive Watcher and list of dead nodes.
Created a set interval which checks if a message hasn't been recieved from a node within ten seconds. If it hasn't then broadcast that the node is not alive and push it to a list of dead nondes.

Created a setup for the docker compose stuff to create a container from the API.

Added some refactorisation to code to send the alive status in the message being broadcast and amended the property names in my leadership election function for readability purposes.