Scale up Containers and Scale Down
Created Set Interval which checks if a node is in the deadnodes list and if so rebuilds another node using the AXIOS Library.

Added API Methods to create and start container and kill and remove container.

Created set interval which creates two new containers between 16:00 and 18:00 and if not within those hours it will scale down by removing last two containers.